+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/home/AndroidHome.kt
@@ -113,9 +113,11 @@ internal fun WrapHome(
     setShowingRestoreSuccess: () -> Unit,
     walletSnapshot: WalletSnapshot?,
 ) {
+    val activity = LocalActivity.current
+
     val focusManager = LocalFocusManager.current
 
-    val activity = LocalActivity.current
+    val walletViewModel by activity.viewModels<WalletViewModel>()
 
     val scope = rememberCoroutineScope()
 
@@ -218,7 +220,9 @@ internal fun WrapHome(
     )
 
     LaunchedEffect(pagerState.currentPage) {
-        if (pagerState.currentPage != HomeScreenIndex.SEND.pageIndex) {
+        if (pagerState.currentPage == HomeScreenIndex.SEND.pageIndex) {
+            walletViewModel.refreshExchangeRateUsd()
+        } else {
             focusManager.clearFocus(true)
         }
     }