+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/send/view/SendView.kt
@@ -4,6 +4,7 @@ package co.electriccoin.zcash.ui.screen.send.view
 
 import androidx.compose.animation.animateContentSize
 import androidx.compose.foundation.ExperimentalFoundationApi
+import androidx.compose.foundation.Image
 import androidx.compose.foundation.ScrollState
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
@@ -12,6 +13,7 @@ import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.requiredSize
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.relocation.BringIntoViewRequester
@@ -32,6 +34,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.focus.FocusDirection
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.ColorFilter
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.layout.onGloballyPositioned
 import androidx.compose.ui.layout.positionInRoot
@@ -47,6 +50,8 @@ import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
 import cash.z.ecc.android.sdk.model.Memo
 import cash.z.ecc.android.sdk.model.MonetarySeparators
 import cash.z.ecc.android.sdk.model.ZecSend
@@ -64,6 +69,7 @@ import co.electriccoin.zcash.ui.common.model.WalletSnapshot
 import co.electriccoin.zcash.ui.common.model.canSpend
 import co.electriccoin.zcash.ui.common.model.spendableBalance
 import co.electriccoin.zcash.ui.common.test.CommonTag
+import co.electriccoin.zcash.ui.common.wallet.ExchangeRateState
 import co.electriccoin.zcash.ui.design.MINIMAL_WEIGHT
 import co.electriccoin.zcash.ui.design.component.AppAlertDialog
 import co.electriccoin.zcash.ui.design.component.BlankBgScaffold
@@ -104,13 +110,19 @@ private fun PreviewSendForm() {
             recipientAddressState = RecipientAddressState("invalid_address", AddressType.Invalid()),
             onRecipientAddressChange = {},
             setAmountState = {},
-            amountState = AmountState.Valid(ZatoshiFixture.ZATOSHI_LONG.toString(), ZatoshiFixture.new()),
+            amountState =
+                AmountState.Valid(
+                    value = ZatoshiFixture.ZATOSHI_LONG.toString(),
+                    fiatValue = "",
+                    zatoshi = ZatoshiFixture.new()
+                ),
             setMemoState = {},
             memoState = MemoState.new("Test message"),
             topAppBarSubTitleState = TopAppBarSubTitleState.None,
             walletSnapshot = WalletSnapshotFixture.new(),
             balanceState = BalanceStateFixture.new(),
             isHideBalances = false,
+            exchangeRateState = ExchangeRateState {}
         )
     }
 }
@@ -135,13 +147,19 @@ private fun SendFormTransparentAddressPreview() {
                 ),
             onRecipientAddressChange = {},
             setAmountState = {},
-            amountState = AmountState.Valid(ZatoshiFixture.ZATOSHI_LONG.toString(), ZatoshiFixture.new()),
+            amountState =
+                AmountState.Valid(
+                    value = ZatoshiFixture.ZATOSHI_LONG.toString(),
+                    fiatValue = "",
+                    zatoshi = ZatoshiFixture.new()
+                ),
             setMemoState = {},
             memoState = MemoState.new("Test message"),
             topAppBarSubTitleState = TopAppBarSubTitleState.None,
             walletSnapshot = WalletSnapshotFixture.new(),
             balanceState = BalanceStateFixture.new(),
             isHideBalances = false,
+            exchangeRateState = ExchangeRateState {}
         )
     }
 }
@@ -170,6 +188,7 @@ fun Send(
     memoState: MemoState,
     topAppBarSubTitleState: TopAppBarSubTitleState,
     walletSnapshot: WalletSnapshot,
+    exchangeRateState: ExchangeRateState,
 ) {
     BlankBgScaffold(topBar = {
         SendTopAppBar(
@@ -202,7 +221,8 @@ fun Send(
                         bottom = paddingValues.calculateBottomPadding() + ZcashTheme.dimens.spacingHuge,
                         start = ZcashTheme.dimens.screenHorizontalSpacingRegular,
                         end = ZcashTheme.dimens.screenHorizontalSpacingRegular
-                    )
+                    ),
+            exchangeRateState = exchangeRateState
         )
     }
 }
@@ -257,6 +277,7 @@ private fun SendMainContent(
     balanceState: BalanceState,
     isHideBalances: Boolean,
     walletSnapshot: WalletSnapshot,
+    exchangeRateState: ExchangeRateState,
     onBack: () -> Unit,
     goBalances: () -> Unit,
     onCreateZecSend: (ZecSend) -> Unit,
@@ -288,7 +309,8 @@ private fun SendMainContent(
         onQrScannerOpen = onQrScannerOpen,
         goBalances = goBalances,
         hasCameraFeature = hasCameraFeature,
-        modifier = modifier
+        modifier = modifier,
+        exchangeRateState = exchangeRateState
     )
 
     if (sendStage is SendStage.SendFailure) {
@@ -312,6 +334,7 @@ private fun SendForm(
     isHideBalances: Boolean,
     walletSnapshot: WalletSnapshot,
     recipientAddressState: RecipientAddressState,
+    exchangeRateState: ExchangeRateState,
     onRecipientAddressChange: (String) -> Unit,
     amountState: AmountState,
     setAmountState: (AmountState) -> Unit,
@@ -363,7 +386,7 @@ private fun SendForm(
         Spacer(Modifier.size(ZcashTheme.dimens.spacingDefault))
 
         SendFormAmountTextField(
-            amountSate = amountState,
+            amountState = amountState,
             imeAction =
                 if (recipientAddressState.type == AddressType.Transparent) {
                     ImeAction.Done
@@ -374,6 +397,7 @@ private fun SendForm(
             monetarySeparators = monetarySeparators,
             setAmountState = setAmountState,
             walletSnapshot = walletSnapshot,
+            exchangeRateState = exchangeRateState
         )
 
         Spacer(Modifier.size(ZcashTheme.dimens.spacingDefault))
@@ -386,7 +410,8 @@ private fun SendForm(
                     recipientAddressState.type is AddressType.Invalid ||
                     (
                         recipientAddressState.type is AddressType.Valid &&
-                            recipientAddressState.type !is AddressType.Transparent
+                            recipientAddressState.type !is AddressType.Transparent &&
+                            recipientAddressState.type !is AddressType.Tex
                     )
             ),
             scrollState = scrollState,
@@ -582,10 +607,11 @@ fun SendFormAddressTextField(
 @Suppress("LongParameterList", "LongMethod")
 @Composable
 fun SendFormAmountTextField(
-    amountSate: AmountState,
+    amountState: AmountState,
     imeAction: ImeAction,
     isTransparentRecipient: Boolean,
     monetarySeparators: MonetarySeparators,
+    exchangeRateState: ExchangeRateState,
     setAmountState: (AmountState) -> Unit,
     walletSnapshot: WalletSnapshot,
 ) {
@@ -596,9 +622,9 @@ fun SendFormAmountTextField(
     val zcashCurrency = ZcashCurrency.getLocalizedName(context)
 
     val amountError =
-        when (amountSate) {
+        when (amountState) {
             is AmountState.Invalid -> {
-                if (amountSate.value.isEmpty()) {
+                if (amountState.value.isEmpty()) {
                     null
                 } else {
                     stringResource(id = R.string.send_amount_invalid)
@@ -606,7 +632,7 @@ fun SendFormAmountTextField(
             }
 
             is AmountState.Valid -> {
-                if (walletSnapshot.spendableBalance() < amountSate.zatoshi) {
+                if (walletSnapshot.spendableBalance() < amountState.zatoshi) {
                     stringResource(id = R.string.send_amount_insufficient_balance)
                 } else {
                     null
@@ -628,47 +654,122 @@ fun SendFormAmountTextField(
 
         Spacer(modifier = Modifier.height(ZcashTheme.dimens.spacingSmall))
 
-        FormTextField(
-            value = amountSate.value,
-            onValueChange = { newValue ->
-                setAmountState(
-                    AmountState.new(
-                        context = context,
-                        value = newValue,
-                        monetarySeparators = monetarySeparators,
-                        isTransparentRecipient = isTransparentRecipient
+        Row {
+            FormTextField(
+                textStyle = ZcashTheme.extendedTypography.textFieldValue.copy(fontSize = 14.sp),
+                value = amountState.value,
+                onValueChange = { newValue ->
+                    setAmountState(
+                        AmountState.newFromZec(
+                            context = context,
+                            value = newValue,
+                            monetarySeparators = monetarySeparators,
+                            isTransparentRecipient = isTransparentRecipient,
+                            fiatValue = amountState.fiatValue,
+                            exchangeRateState = exchangeRateState
+                        )
                     )
+                },
+                modifier = Modifier.weight(1f),
+                error = amountError,
+                placeholder = {
+                    Text(
+                        text =
+                            stringResource(
+                                id = R.string.send_amount_hint,
+                                zcashCurrency
+                            ),
+                        style = ZcashTheme.extendedTypography.textFieldHint,
+                        color = ZcashTheme.colors.textFieldHint
+                    )
+                },
+                keyboardOptions =
+                    KeyboardOptions(
+                        keyboardType = KeyboardType.Number,
+                        imeAction = imeAction
+                    ),
+                keyboardActions =
+                    KeyboardActions(
+                        onDone = {
+                            focusManager.clearFocus(true)
+                        },
+                        onNext = {
+                            focusManager.moveFocus(FocusDirection.Right)
+                        }
+                    ),
+                bringIntoViewRequester = bringIntoViewRequester,
+                leadingIcon = {
+                    Image(
+                        modifier = Modifier.requiredSize(7.dp, 13.dp),
+                        painter = painterResource(R.drawable.ic_send_zashi),
+                        contentDescription = "",
+                        colorFilter = ColorFilter.tint(color = ZcashTheme.colors.secondaryColor),
+                    )
+                }
+            )
+
+            if (!exchangeRateState.isStale || exchangeRateState.isLoading) {
+                Spacer(modifier = Modifier.width(ZcashTheme.dimens.spacingMin))
+                Image(
+                    modifier = Modifier.padding(top = 24.dp),
+                    painter = painterResource(id = R.drawable.ic_send_convert),
+                    contentDescription = "",
+                    colorFilter = ColorFilter.tint(color = ZcashTheme.colors.secondaryColor),
                 )
-            },
-            modifier = Modifier.fillMaxWidth(),
-            error = amountError,
-            placeholder = {
-                Text(
-                    text =
-                        stringResource(
-                            id = R.string.send_amount_hint,
-                            zcashCurrency
-                        ),
-                    style = ZcashTheme.extendedTypography.textFieldHint,
-                    color = ZcashTheme.colors.textFieldHint
-                )
-            },
-            keyboardOptions =
-                KeyboardOptions(
-                    keyboardType = KeyboardType.Number,
-                    imeAction = imeAction
-                ),
-            keyboardActions =
-                KeyboardActions(
-                    onDone = {
-                        focusManager.clearFocus(true)
+                Spacer(modifier = Modifier.width(ZcashTheme.dimens.spacingMin))
+                FormTextField(
+                    textStyle = ZcashTheme.extendedTypography.textFieldValue.copy(fontSize = 14.sp),
+                    value = amountState.fiatValue,
+                    onValueChange = { newValue ->
+                        setAmountState(
+                            AmountState.newFromFiat(
+                                context = context,
+                                value = amountState.value,
+                                monetarySeparators = monetarySeparators,
+                                isTransparentRecipient = isTransparentRecipient,
+                                fiatValue = newValue,
+                                exchangeRateState = exchangeRateState
+                            )
+                        )
                     },
-                    onNext = {
-                        focusManager.moveFocus(FocusDirection.Down)
+                    modifier = Modifier.weight(1f),
+                    placeholder = {
+                        Text(
+                            text =
+                                stringResource(
+                                    id = R.string.send_usd_amount_hint,
+                                    zcashCurrency
+                                ),
+                            style = ZcashTheme.extendedTypography.textFieldHint,
+                            color = ZcashTheme.colors.textFieldHint
+                        )
+                    },
+                    keyboardOptions =
+                        KeyboardOptions(
+                            keyboardType = KeyboardType.Number,
+                            imeAction = imeAction
+                        ),
+                    keyboardActions =
+                        KeyboardActions(
+                            onDone = {
+                                focusManager.clearFocus(true)
+                            },
+                            onNext = {
+                                focusManager.moveFocus(FocusDirection.Down)
+                            }
+                        ),
+                    bringIntoViewRequester = bringIntoViewRequester,
+                    leadingIcon = {
+                        Image(
+                            modifier = Modifier.requiredSize(7.dp, 13.dp),
+                            painter = painterResource(R.drawable.ic_usd),
+                            contentDescription = "",
+                            colorFilter = ColorFilter.tint(color = ZcashTheme.colors.secondaryColor),
+                        )
                     }
-                ),
-            bringIntoViewRequester = bringIntoViewRequester,
-        )
+                )
+            }
+        }
     }
 }