+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/screen/sendconfirmation/AndroidSendConfirmation.kt
@@ -29,6 +29,7 @@ import co.electriccoin.zcash.ui.R
 import co.electriccoin.zcash.ui.common.model.TopAppBarSubTitleState
 import co.electriccoin.zcash.ui.common.viewmodel.AuthenticationViewModel
 import co.electriccoin.zcash.ui.common.viewmodel.WalletViewModel
+import co.electriccoin.zcash.ui.common.wallet.ExchangeRateState
 import co.electriccoin.zcash.ui.design.component.CircularScreenProgressIndicator
 import co.electriccoin.zcash.ui.screen.authentication.AuthenticationUseCase
 import co.electriccoin.zcash.ui.screen.authentication.WrapAuthentication
@@ -74,6 +75,8 @@ internal fun MainActivity.WrapSendConfirmation(
 
     val walletState = walletViewModel.walletStateInformation.collectAsStateWithLifecycle().value
 
+    val exchangeRateState = walletViewModel.exchangeRateUsd.collectAsStateWithLifecycle().value
+
     WrapSendConfirmation(
         activity = this,
         arguments = arguments,
@@ -87,6 +90,7 @@ internal fun MainActivity.WrapSendConfirmation(
         supportMessage = supportMessage,
         synchronizer = synchronizer,
         topAppBarSubTitleState = walletState,
+        exchangeRateState = exchangeRateState,
     )
 }
 
@@ -97,6 +101,7 @@ internal fun WrapSendConfirmation(
     activity: MainActivity,
     arguments: SendConfirmationArguments,
     authenticationViewModel: AuthenticationViewModel,
+    exchangeRateState: ExchangeRateState,
     createTransactionsViewModel: CreateTransactionsViewModel,
     goBack: (clearForm: Boolean) -> Unit,
     goHome: () -> Unit,
@@ -206,6 +211,7 @@ internal fun WrapSendConfirmation(
                 }
             },
             topAppBarSubTitleState = topAppBarSubTitleState,
+            exchangeRate = exchangeRateState
         )
 
         if (sendFundsAuthentication.value) {